requirements.txt
streamlit
python-dateutil
groq
ollama
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
langchain
langchain_ollama
dotenv
dateparser
langchain_community
PyPDF2

# --- NUEVAS DEPENDENCIAS PARA LA ENTREGA ---
openai
crewai
langchain-openai
chromadb
langchain-chroma
pypdf